Allocator fixes, and some missing includes.
The allocator was a bit misleading in that when new allocations were done with
a specified size limit the allocator would take a small chunk of the space
available internally giving the caller a false sense of available storage.

There are also a few includes that were left out of the AddressMapper.  Those
have now been included.