GOT PONG! (fixed packet less than minimum size)
Fixed small packets smaller than minimum size

Ethernet payload must be at least of 46 bytes.

cjdns has some 40 bytes small messages

So if we detect a received size of 46 ('cause on sending the packet was
padded), we magically say it's 40 bytes.

Still need to get it right, tough. but:

        WE GOT PONG!

            \o/

Also:
* don't pad ETH_ALEN on sent message (already done)
* removed assert on addrLen == sizeof(struct sockaddr_ll)
    => we get addrLen == 18 (maybe two null bytes)
* wrote ETHInterface example config in cjdroute.c
* nicer syntax for baseAddr initialization :)
* padding: 0 -> 16 (because assert failed)
* using "addr" instead of "sll"
* removed some debug code