cjdns.sh: Add status command
The check command is the closest available to check whether cjdns is
running. However, it has the side effect of starting cjdns if it is not
running.

Add the "status" command to report whether cjdns is running. This is
similar to a lot of init.d scripts for services. If cjdns is running, an
exit code of 0 will be returned. Otherwise an exit code of 1 is
returned.

	$ ./scripts/cjdns.sh status
	* cjdns is running