Support for endpoints with dynamic IP, via dyndns
Also fixes a truncation thing that was making the reported unresponsive
node's public key truncated. (52 = 32 * 13 / 8 ?? how do I base32?).
The idea is not to specify any connectTo in cjdroute.conf, but to list
the endpoints in dynamicEndpoints and run it, leaving it running to
catch any that go unresponsive (possibly because their IP changed)