Remove fork in script to capture exit status
The purpose of this commit is to improve user output when invoking
start(). It removes the fork when starting cjdns so as to capture the
output status. The fork has been irrelevant since the time angel and
core were forked off of the main invocation.

In addition, it prints a brief error message explaining the output to
the user, assuming the startup fails.

    "Failed to start. (CJDNS already running?)"