[HARDENING] Add code to specifically prevent decrypted messages from leaking in error packets. Previously, Ducttape could return an error which would cause the SwitchCore to send back an error packet to the original sender of the message so conceivably there could have been a security issue where some packets were decrypted and sent back over the wire.
In practice Ducttape only actually returned an error if there was an AUTHENTICATION error
and Error_AUTHENTICATION is only returned by CryptoAuth itself and only before the content
is decrypted so this was not actually an issue but that was only because of luck.

This patch changes SwitchCore to never send anything over the wire except that which came
in, no matter what the return value is from the upper layer so this class of bug is not possible.