Make errors in nacl build get reported to console
The "stderr" in the callback from running the compiler to build nacl was really getting the standard out data, while the standard error data was being dropped. This commit adds an extra parameter to hold the standard out data, causing "stderr" to actually get the standard error data as it is supposed to, allowing problems with the nacl build to be debugged.