Added option to ./do to download node.js source and compile in-tree
While the ./do script is able to download binaries for platforms, this
addition adds a function buildNode() to ./do, which downloads the hefty
source code package, opens it, builds it into the location where
hasOkNode() checks for it, and then the program proceeds with building
cjdns.

Don't mind the "deleted ./update" change on the change log, this branch
was cloned from my master, which is where my ./update feature is
currently under review. ./update was deleted in this branch to conform
with upstream.