client: exit foreground process when core dies
Many init scripts run cjdroute in foreground, so they can easily
send signals and fetch logs. They can however not monitor cjdroute,
because the foreground process doesn't keep track of the forked
core process. Thus recovery in case of crashes requires a manual
restart of cjdroute.

With this patch, once the core process exits, the foreground
process prints the exit code and the signal which caused the
exit. It then exits, too, and the init script can restart it.