android: make Pipe_PREFIX configurable at build time
In the context of an Android app, the prefix depends on
the name of the app, because the app's directory is the
only writable path.

Example: /data/data/berlin.meshnet.cjdns/cjdns_pipe_

The other writable path (SD card) is not an option,
because it's usually FAT32/VFAT which doesn't support
named pipes.